Move recyclerview functionality into a use case
