+++ room-kit/src/main/java/live/hms/roomkit/ui/meeting/participants/ParticipantsTabFragment.kt
@@ -34,17 +34,12 @@ class ParticipantsTabFragment(val dismissFragment: () -> Unit) : Fragment() {
     private val TAG = "ParticipantsFragment"
     private var binding by viewLifecycle<LayoutParticipantsMergeBinding>()
     private var alertDialog: AlertDialog? = null
-    val adapter = GroupieAdapter()
-    private lateinit var handRaisedKey :String
-    private var filterText : String? = null
-    private fun isSearching() = !filterText.isNullOrEmpty()
-    private val expandedGroups = mutableMapOf<String,Boolean>()
-
     private val meetingViewModel: MeetingViewModel by activityViewModels {
         MeetingViewModelFactory(
             requireActivity().application
         )
     }
+    private val participantsUseCase by lazy { ParticipantsUseCase(meetingViewModel) { meetingViewModel.peers } }
 
 
     override fun onCreateView(
@@ -53,81 +48,8 @@ class ParticipantsTabFragment(val dismissFragment: () -> Unit) : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         binding = LayoutParticipantsMergeBinding.inflate(inflater, container, false)
-        handRaisedKey = requireContext().resources.getString(R.string.hand_raised_group)
         return binding.root
     }
-
-    // This is only suspending so it can run in the background
-    private suspend fun updateParticipantsAdapter(peerList: List<HMSPeer>) {
-        // Don't throw away results when it's searching
-        //  ideally this should be replaced with just updating the
-        //  peers but still with the search query.
-
-        // Group people by roles.
-        val groupedPeers : Map<String, List<HMSPeer>> = peerList.groupBy {
-            if(CustomPeerMetadata.fromJson(it.metadata)?.isHandRaised == true)
-                handRaisedKey
-            else
-                it.hmsRole.name
-        }
-
-
-        val canChangeRole = meetingViewModel.isAllowedToChangeRole()
-        val canMutePeers = meetingViewModel.isAllowedToMutePeers()
-        val canRemovePeers = meetingViewModel.isAllowedToRemovePeers()
-        val localPeer = meetingViewModel.hmsSDK.getLocalPeer()!!
-
-        val groups = mutableListOf<ExpandableGroup>()
-        // Keep hand raised on top.
-        if(groupedPeers[handRaisedKey] != null) {
-            groups.add(keyToGroup(handRaisedKey, groupedPeers, canChangeRole, canMutePeers, canRemovePeers, localPeer))
-        }
-
-        groups.addAll(groupedPeers.keys.filterNot { it == handRaisedKey }.map { key ->
-            keyToGroup(key, groupedPeers, canChangeRole, canMutePeers, canRemovePeers, localPeer)
-        })
-
-        adapter.update(groups)
-    }
-
-
-    private fun expandedGroups( rolename : String, expanded : Boolean) {
-        expandedGroups[rolename] = expanded
-    }
-
-    private fun keyToGroup(
-        key: String,
-        groupedPeers: Map<String, List<HMSPeer>>,
-        canChangeRole: Boolean,
-        canMutePeers: Boolean,
-        canRemovePeers: Boolean,
-        localPeer : HMSLocalPeer
-    ) : ExpandableGroup =
-        ExpandableGroup(ParticipantHeaderItem(key, groupedPeers[key]?.size, ::expandedGroups))
-            .apply {
-                addAll(groupedPeers[key]?.map {
-                    ParticipantItem(
-                        it,
-                        localPeer,
-                        meetingViewModel::togglePeerMute,
-                        ::changePeerRole,
-                        canChangeRole,
-                        canMutePeers,
-                        canRemovePeers,
-                        meetingViewModel.prebuiltInfoContainer,
-                        meetingViewModel.participantPreviousRoleChangeUseCase,
-                        meetingViewModel::requestPeerLeave,
-                        meetingViewModel.activeSpeakers
-                    )
-                }!!)
-                // If the group was expanded, open it again.
-                if(expandedGroups[key] == true || expandedGroups[key] == null){
-                    onToggleExpanded()
-                }
-            }
-
-    private fun changePeerRole(remotePeerId : String, toRole : String, force : Boolean) =
-        meetingViewModel.changeRole(remotePeerId, toRole, force)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -142,59 +64,19 @@ class ParticipantsTabFragment(val dismissFragment: () -> Unit) : Fragment() {
     }
 
     private fun initViews() {
-        binding.recyclerView.apply {
-            layoutManager = LinearLayoutManager(requireContext())
-            itemAnimator = null
-            addItemDecoration(
-                // Border bright
-                HeaderItemDecoration(
-                    getColorOrDefault(
-                        HMSPrebuiltTheme.getColours()?.borderBright,
-                        HMSPrebuiltTheme.getDefaults().border_bright
-                    ),
-                    0,
-                    16f,
-                    24f,
-                    R.layout.participant_header_item
-                )
-            )
-            val divider = DividerItemDecoration(requireContext(), RecyclerView.VERTICAL).apply {
-                setDrawable(resources.getDrawable(R.drawable.participants_divider)!!)
-            }
-            addItemDecoration(divider)
-        }
+        participantsUseCase.initRecyclerView(binding.recyclerView)
 
         // Search disables conventional updates.
-        binding.textInputSearch.apply {
-            addTextChangedListener { text ->
-                lifecycleScope.launch {
-                    filterText = text.toString()
-                    updateParticipantsAdapter(getSearchFilteredPeersIfNeeded())
-                }
-            }
-        }
+        participantsUseCase.initSearchView(binding.textInputSearch, lifecycleScope)
     }
 
-    private fun getSearchFilteredPeersIfNeeded() : List<HMSPeer> {
-        val text = filterText
-
-        return if (!isSearching())
-            meetingViewModel.peers
-        else
-            meetingViewModel.peers.filter {
-                text.isNullOrEmpty() || it.name.contains(
-                    text.toString(),
-                    true
-                )
-            }
-    }
     @SuppressLint("SetTextI18n")
     private fun initViewModels() {
-        binding.recyclerView.adapter = adapter
+        binding.recyclerView.adapter = participantsUseCase.adapter
         // Initial updating of views
         meetingViewModel.participantPeerUpdate.observe(viewLifecycleOwner) {
             lifecycleScope.launch {
-                updateParticipantsAdapter(getSearchFilteredPeersIfNeeded())
+                participantsUseCase.updateParticipantsAdapter(meetingViewModel.peers)
             }
         }
         meetingViewModel.peerCount.observe(viewLifecycleOwner,::updateParticipantCount)
